% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-wos.R
\name{query_wos}
\alias{query_wos}
\title{Query the Web of Science}
\usage{
query_wos(query, edition = c("SCI", "SSCI", "AHCI", "ISTP", "ISSHP", "BSCI",
  "BHCI", "IC", "CCR", "ESCI"), email = NULL, sid = auth())
}
\arguments{
\item{query}{Query string}

\item{edition}{Web of Science editions to query. Possible values are listed
online \href{http://ipscience-help.thomsonreuters.com/wosWebServicesLite/dbEditionsOptionsGroup/databaseEditionsWos.html}{here}.}

\item{email}{Your email. Include this so the good people at Clarivate
Analytics know who's accessing their service. The default value of \code{NULL}
means no email will be included in the HTTP request.}

\item{sid}{Session identifier (SID). The default setting is to get a fresh
SID each time you query the WoS, via a call to \code{\link{auth}}. However,
you should try to reuse SID values over multiple queries so that you don't
run into the throttling limits placed on new sessions.}
}
\value{
An object of class \code{query_result}. This object has the number
of publications that are returned by your query, as well as all the info
you'll need to download the data.
}
\description{
Query the Web of Science
}
\examples{
\dontrun{

query_wos(query = "TS = (dog welfare) AND PY = (1990-2007)")
}

}
